src/main/java/com/pedrogallardo/cli/Item.java
package com.pedrogallardo.cli;

import java.util.List;

public class Item {
    private String name;
    private List<Item> items;

    public Item(String name, List<Item> items) {
        this.name = name;
        this.items = items;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public List<Item> getItems() {
        return items;
    }

    public void setItems(List<Item> items) {
        this.items = items;
    }
}
